requirements.txt
Django==2.2.13
django-cors-headers==3.4.0
djangorestframework==3.11.0
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
